Fix more mistakes

+++ src/core/scene.cpp
@@ -196,12 +196,12 @@ void Scene::removeScript(unsigned int id)
 void Scene::triggerUpdate()
 {
     auto now = std::chrono::system_clock::now();
-    std::chrono::duration<float> delta = now - _lastTime;
+    std::chrono::duration<double> delta = now - _lastTime;
     _lastTime = now;
 
     for (auto it = _scripts.begin(); it != _scripts.end(); it++)
     {
-        it->second->update(delta.count());
+        it->second->update((float)(delta.count()));
     }
 }
